Fix "bug" where categories would be discarded if not in the exact, correct order.
To be honest, I'm not sure if this is a bug or not (hence the quotes
above) but it's really not nice to dump stuff from the category tree
just because children don't appear *immediately* after the parent.

This stops that happening, as well as simplifying the code a little.